Fix formula for Mu.
The formula for the Cm-based derivative Mu is updated. The force (Zu and Xu) and torque (Mu) derivatives with respect to forward speed (u) depend not only on the change in coefficients (Cl, Cd and Cm) but also on the change to the speed itself. The point is that u*u*Cm will depend on changes to Cm and also u. In most cases this fix will not matter; usually the scenario for the Mu calculation is one where the aircraft is pitch stable (Cm = 0) and only the change to Cm will matter then. In some scenarios, viz. when a rocket is steered by reaction wheels, the updated formula is more correct.