Cosmetic updates to the in-editor display of stability derivatives.
Reorder stability derivatives. Let us try it and see if it is better or worse.

Reintroduce the "b/(2u)" and "MAC/(2u)" factors used to scale static stability derivatives with respect to rotations, but only in the StabilityDerivLinearSim class. The purpose is backward compability. The scaling will (most likely) be disabled in the next commit. The idea is that this way it is easy to reenable the classic scaling at will.